backend/requirements.txt
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
requests==2.25.1
SQLAlchemy==1.4.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
asyncpg==0.25.0

bcrypt==3.2.0
passlib==1.7.4
python-multipart==0.0.5

cryptography==3.4.7
ecdsa==0.14.1
pyasn1==0.4.8
python-jose==3.2.0
rsa==4.7.2